+++ travail_de_groupe/jeu_appren_par_renfo/src/ball.h
@@ -29,7 +29,7 @@ extern int trajectoireAntoine[NUMBERPOINT_TRAJEC][2];
 void initBall();
 float defineAngle(int, int, int, int);
 
-int lagrangeInterpolation(float, int, int, int, int, int, int);
+float lagrangeInterpolation(float, int, int, int, int, int, int);
 void calculTrajectoireAntoine2(int, int, int, int, int, int);
 
 #endif
